refactor(game): split builder into focused modules
